refactor: split next_ver module
